Add SPI for RequestLogFactory
