+++ dropwizard-jetty/src/main/resources/META-INF/services/io.dropwizard.jetty.RequestLogFactory
@@ -1,0 +1,1 @@
+io.dropwizard.jetty.Slf4jRequestLogFactory
